requirements.txt
Flask
Flask-RESTful
Flask-Cors
Flask-Migrate
flask-sqlalchemy
python-dotenv
Flask-JWT-Extended
azure-identity
azure-search-documents
langchain==0.3.27
langchain-openai
langchain-community
gunicorn
pypdf
html2text
spacy
markdown
beautifulsoup4
pymysql
Flask-Limiter